migration update

+++ database/migrations/2018_10_29_223723_create_providers_table.php
@@ -15,7 +15,7 @@ class CreateProvidersTable extends Migration
     {
         Schema::create('providers', function (Blueprint $table) {
             $table->increments('id');
-            $table->integer('provider_id');
+            $table->bigInteger('provider_id');
             $table->string('provider');
             $table->unsignedInteger('user_id');
             $table->timestamps();
